+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 3.2.0
 * Deprecate `MasterpassCard` payment method
 * Deprecate `AmexExpressCheckoutCard` payment method
 * Deprecate `deviceSessionId` and `fraudMerchantId` in `CreditCardRequest`, `CustomerRequest`, `PaymentMethodRequest`, and `TransactionRequest` classes